sfa/web/requirements.txt
Django==2.1.4
pinax-templates==2.0.1
django-user-accounts==2.0.3
pinax-eventlog==2.0.3
pinax-webanalytics==4.0.2
django-twilio==0.9.3
django-import-export==1.1.0
django-constance==2.3.1
django-picklefield==2.0
django-debug-toolbar==1.11
phonenumbers==8.10.3